requirements.txt
keras
numpy
Pillow
tensorflow